chefmate_backend/requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings
motor==3.7.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
argon2-cffi==25.1.0
websockets==16.0
email-validator==2.3.0
httptools==0.7.1
uvloop
google-generativeai
